refactor: 쉐이더 핫리로드 리팩토링 완료
