+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Renderer/Renderer.cpp
@@ -17,60 +17,6 @@
 #include <UObject/Casts.h>
 #include "GameFrameWork/Actor.h"
 
-DWORD WINAPI DirectoryChangeWatcher(LPVOID lpParam)
-{
-    FWatchParams* pParams = reinterpret_cast<FWatchParams*>(lpParam);
-    
-    FDXDShaderManager* pShaderManager = pParams->pShaderManager;
-    std::wstring directory = pParams->Directory;
-    delete pParams;
-
-    HANDLE hDir = CreateFileW(
-        directory.c_str(),
-        FILE_LIST_DIRECTORY,
-        FILE_SHARE_READ | FILE_SHARE_WRITE | FILE_SHARE_DELETE,
-        NULL,
-        OPEN_EXISTING,
-        FILE_FLAG_BACKUP_SEMANTICS,
-        NULL);
-
-    if (hDir == INVALID_HANDLE_VALUE)
-    {
-        OutputDebugStringA("디렉토리 감시 핸들 생성 실패\n");
-        return 1;
-    }
-
-    const DWORD bufferSize = 4096;
-    BYTE buffer[bufferSize];
-    DWORD bytesReturned = 0;
-
-    while (true)
-    {
-       BOOL success = ReadDirectoryChangesW(
-            hDir,
-            buffer,
-            bufferSize,
-            TRUE,
-            FILE_NOTIFY_CHANGE_LAST_WRITE,
-            &bytesReturned,
-            NULL, 
-            NULL);
-
-        if (success)
-        {
-            pShaderManager->CheckAndReloadShaders();
-        }
-        else
-        {
-            OutputDebugStringA("ReadDirectoryChangesW 실패\n");
-        }
-        Sleep(50);
-    }
-
-    CloseHandle(hDir);
-    return 0;
-}
-
 
 //------------------------------------------------------------------------------
 // 초기화 및 해제 관련 함수
@@ -81,6 +27,8 @@ void FRenderer::Initialize(FGraphicsDevice* InGraphics, FDXDBufferManager* InBuf
     BufferManager = InBufferManager;
 
     ShaderManager = new FDXDShaderManager(Graphics->Device, Graphics);
+    ShaderHotReload = new FShaderHotReload(ShaderManager);
+
     StaticMeshRenderPass = new FStaticMeshRenderPass();
     BillboardRenderPass = new FBillboardRenderPass();
     GizmoRenderPass = new FGizmoRenderPass();
@@ -98,15 +46,10 @@ void FRenderer::Initialize(FGraphicsDevice* InGraphics, FDXDBufferManager* InBuf
     DepthBufferDebugPass->Initialize(BufferManager, Graphics, ShaderManager);
     FogRenderPass->Initialize(BufferManager, Graphics, ShaderManager);
     LightCullPass->Initialize(BufferManager, Graphics, ShaderManager);
+    
+    ShaderHotReload->CreateShaderHotReloadThread();
 
     CreateConstantBuffers();
-
-    FWatchParams* pParams = new FWatchParams();
-    pParams->pShaderManager = ShaderManager;  // 이미 생성된 FDXDShaderManager 인스턴스
-    pParams->Directory = L"Shaders";
-
-
-    HANDLE hThread = CreateThread( nullptr, 0, DirectoryChangeWatcher, pParams, 0, nullptr);
 }
 
 void FRenderer::Release()
@@ -153,7 +96,7 @@ void FRenderer::CreateConstantBuffers()
 
     UINT FogConstantBufferSize = sizeof(FFogConstants);
     BufferManager->CreateBufferGeneric<FFogConstants>("FFogConstants", nullptr, FogConstantBufferSize, D3D11_BIND_CONSTANT_BUFFER, D3D11_USAGE_DYNAMIC, D3D11_CPU_ACCESS_WRITE);
-    
+
     UINT FPrimitiveCountsSize = sizeof(FPrimitiveCounts);
     BufferManager->CreateBufferGeneric<FPrimitiveCounts>("FPrimitiveCounts", nullptr, FPrimitiveCountsSize, D3D11_BIND_CONSTANT_BUFFER, D3D11_USAGE_DYNAMIC, D3D11_CPU_ACCESS_WRITE);
 
@@ -200,7 +143,7 @@ void FRenderer::Render(const std::shared_ptr<FEditorViewportClient>& ActiveViewp
     StaticMeshRenderPass->SwitchShaderLightingMode(ActiveViewport->GetViewMode());
     StaticMeshRenderPass->Render(ActiveViewport);
     BillboardRenderPass->Render(ActiveViewport);
-    
+
 
     if (IsSceneDepth)
     {
@@ -210,7 +153,7 @@ void FRenderer::Render(const std::shared_ptr<FEditorViewportClient>& ActiveViewp
     if (!IsSceneDepth)
     {
         DepthBufferDebugPass->UpdateDepthBufferSRV();
-        
+
         FogRenderPass->RenderFog(ActiveViewport, Graphics->DepthBufferSRV);
     }
