+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Windows/D3D11RHI/DXDShaderManager.cpp
@@ -1,146 +1,9 @@
 #include "DXDShaderManager.h"
 #include "D3D11RHI/GraphicDevice.h"
-#include <sstream>
-#include <functional>
+#include "HotReload/ShaderHotReload.h"
+#include "HotReload/ShaderHashUtils.h"
 #include <d3d11.h>
 #include <d3dcompiler.h>
-#include <Windows.h>
-#include <fstream>
-#include <sstream>
-
-// 익명 네임스페이스 내에 유틸리티 함수들을 정의
-namespace {
-    // 파일 내용을 읽어 std::hash<std::string>()를 사용하여 해시값 계산
-    std::size_t ComputeFileHash(const std::wstring& filename)
-    {
-        std::ifstream file(filename, std::ios::binary);
-        if (!file)
-            return 0;
-        std::ostringstream contents;
-        contents << file.rdbuf();
-        std::string fileContent = contents.str();
-        return std::hash<std::string>()(fileContent);
-    }
-
-    // 셰이더 컴파일 공통 함수
-    HRESULT CompileShader(const std::wstring& FileName,
-        const D3D_SHADER_MACRO* Defines,
-        const std::string& EntryPoint,
-        const std::string& Target,
-        ID3DBlob** OutBlob)
-    {
-        UINT shaderFlags = D3DCOMPILE_ENABLE_STRICTNESS;
-#ifdef _DEBUG
-        shaderFlags |= D3DCOMPILE_DEBUG | D3DCOMPILE_SKIP_OPTIMIZATION;
-#endif
-        return D3DCompileFromFile(FileName.c_str(), Defines, D3D_COMPILE_STANDARD_FILE_INCLUDE,
-            EntryPoint.c_str(), Target.c_str(), shaderFlags, 0, OutBlob, nullptr);
-    }
-
-    HRESULT CompileShaderWithTempFile(const std::wstring& originalFile, const D3D_SHADER_MACRO* Defines, const std::string& EntryPoint,
-        const std::string& ShaderModel, UINT shaderFlags, ID3DBlob** blobOut)
-    {
-        std::wstring tempFile;
-        size_t pos = originalFile.find_last_of(L"\\/");
-        if (pos != std::wstring::npos)
-        {
-            tempFile = originalFile.substr(0, pos + 1) + L"temp_" + originalFile.substr(pos + 1);
-        }
-        else
-        {
-            tempFile = L"temp_" + originalFile;
-        }
-
-        const int maxRetry = 30;
-        int retry = 0;
-        BOOL copySuccess = FALSE;
-        while (retry < maxRetry)
-        {
-            copySuccess = CopyFileW(originalFile.c_str(), tempFile.c_str(), FALSE);
-
-            if (copySuccess)
-                break;
-
-            retry++;
-
-            Sleep(10); // 10ms 지연 후 재시도
-        }
-
-        if (!copySuccess)
-        {
-            OutputDebugStringA("임시 파일 복사 실패\n");
-            return E_FAIL;
-        }
-
-        ID3DBlob* errorBlob = nullptr;
-        HRESULT hr = D3DCompileFromFile(tempFile.c_str(), Defines, D3D_COMPILE_STANDARD_FILE_INCLUDE, EntryPoint.c_str(), ShaderModel.c_str(), shaderFlags, 0, blobOut, &errorBlob);
-        DeleteFileW(tempFile.c_str());
-
-
-        if (FAILED(hr))
-        {
-            if (errorBlob)
-            {
-                OutputDebugStringA((char*)errorBlob->GetBufferPointer());
-                errorBlob->Release();
-            }
-            return hr;
-        }
-
-        return hr;
-    }
-
-    HRESULT ReloadShader(ID3D11Device* Device,
-        const std::wstring& FileName,
-        const D3D_SHADER_MACRO* Defines,
-        const std::string& EntryPoint,
-        const std::string& Target,
-        void** NewShader)
-    {
-        if (Device == nullptr)
-        {
-            OutputDebugStringA("Device 포인터가 NULL입니다.\n");
-            return E_INVALIDARG;
-        }
-        if (NewShader == nullptr)
-        {
-            OutputDebugStringA("NewShader 포인터가 NULL입니다.\n");
-            return E_INVALIDARG;
-        }
-
-        ID3DBlob* shaderBlob = nullptr;
-        HRESULT hr = CompileShader(FileName, Defines, EntryPoint, Target, &shaderBlob);
-        if (FAILED(hr))
-        {
-            if (shaderBlob)
-            {
-                OutputDebugStringA("Shader 컴파일에 실패하였습니다.\n");
-                shaderBlob->Release();
-            }
-            return hr;
-        }
-
-        if (Target == "vs_5_0")
-        {
-            hr = Device->CreateVertexShader(shaderBlob->GetBufferPointer(), shaderBlob->GetBufferSize(), nullptr,
-                reinterpret_cast<ID3D11VertexShader**>(NewShader));
-        }
-        else if (Target == "ps_5_0")
-        {
-            hr = Device->CreatePixelShader(shaderBlob->GetBufferPointer(), shaderBlob->GetBufferSize(), nullptr,
-                reinterpret_cast<ID3D11PixelShader**>(NewShader));
-        }
-        else
-        {
-            OutputDebugStringA("알 수 없는 타겟입니다: ");
-            OutputDebugStringA(Target.c_str());
-            OutputDebugStringA("\n");
-            hr = E_INVALIDARG;
-        }
-        shaderBlob->Release();
-        return hr;
-    }
-}
 
 // --------------------------------------------------------------------------
 // FDXDShaderManager 클래스 구현
@@ -174,24 +37,6 @@ void FDXDShaderManager::ReleaseAllShader()
         }
     }
     PixelShaders.Empty();
-}
-
-size_t FDXDShaderManager::CalculateShaderHashKey(const std::wstring& FileName, const std::string& EntryPoint, const D3D_SHADER_MACRO* Defines)
-{
-    std::wstringstream ss;
-    ss << FileName << L"_" << std::wstring(EntryPoint.begin(), EntryPoint.end());
-
-    if (Defines == nullptr)
-        return std::hash<std::wstring>()(ss.str());
-
-    for (int i = 0; Defines[i].Name != nullptr; ++i)
-    {
-        ss << L"_" << std::wstring(Defines[i].Name, Defines[i].Name + strlen(Defines[i].Name));
-        ss << L"_" << std::wstring(Defines[i].Definition, Defines[i].Definition + strlen(Defines[i].Definition));
-    }
-
-    std::wstring combinedStr = ss.str();
-    return std::hash<std::wstring>()(combinedStr);
 }
 
 D3D_SHADER_MACRO* FDXDShaderManager::GetShaderMacro(EViewModeIndex ViewMode)
@@ -257,7 +102,7 @@ HRESULT FDXDShaderManager::AddVertexShader(const std::wstring& Key, const std::w
 HRESULT FDXDShaderManager::AddVertexShader(const std::wstring& FileName, const std::string& EntryPoint, EViewModeIndex ViewMode, size_t& OutShaderKey)
 {
     const D3D_SHADER_MACRO* Defines = GetShaderMacro(ViewMode);
-    size_t shaderKey = CalculateShaderHashKey(FileName, EntryPoint, Defines);
+    size_t shaderKey = ShaderHashUtils::ComputeHashKey(ShaderCompileInfo(FileName, EntryPoint, Defines));
 
     OutShaderKey = shaderKey;
 
@@ -301,7 +146,7 @@ HRESULT FDXDShaderManager::AddVertexShader(const std::wstring& FileName, const s
     VertexShaders[shaderKey] = NewVertexShader;
 
     // 해시값 계산하여 저장
-    std::size_t currentHash = ComputeFileHash(FileName);
+    std::size_t currentHash = ShaderHashUtils::ComputeFileHash(FileName);
     FShaderReloadInfo reloadInfo;
     reloadInfo.FileName = FileName;
     reloadInfo.EntryPoint = EntryPoint;
@@ -318,7 +163,7 @@ HRESULT FDXDShaderManager::AddVertexShader(const std::wstring& FileName, const s
 HRESULT FDXDShaderManager::AddPixelShader(const std::wstring& FileName, const std::string& EntryPoint, EViewModeIndex ViewMode, size_t& OutShaderKey)
 {
     const D3D_SHADER_MACRO* Defines = GetShaderMacro(ViewMode);
-    size_t shaderKey = CalculateShaderHashKey(FileName, EntryPoint, Defines);
+    size_t shaderKey = ShaderHashUtils::ComputeHashKey(ShaderCompileInfo(FileName, EntryPoint, Defines));
     OutShaderKey = shaderKey;
 
     if (PixelShaders.Contains(shaderKey))
@@ -351,7 +196,7 @@ HRESULT FDXDShaderManager::AddPixelShader(const std::wstring& FileName, const st
     if (FAILED(hr))
         return hr;
 
-    std::size_t currentHash = ComputeFileHash(FileName);
+    std::size_t currentHash = ShaderHashUtils::ComputeFileHash(FileName);
     FShaderReloadInfo reloadInfo;
     reloadInfo.FileName = FileName;
     reloadInfo.EntryPoint = EntryPoint;
@@ -397,7 +242,7 @@ HRESULT FDXDShaderManager::AddVertexShaderAndInputLayout(const std::wstring& Fil
     ID3DBlob* VertexShaderCSO = nullptr;
     ID3DBlob* ErrorBlob = nullptr;
     const D3D_SHADER_MACRO* Defines = GetShaderMacro(ViewMode);
-    size_t shaderKey = CalculateShaderHashKey(FileName, EntryPoint, Defines);
+    size_t shaderKey = ShaderHashUtils::ComputeHashKey(ShaderCompileInfo(FileName, EntryPoint, Defines));
 
     OutShaderKey = shaderKey;
 
@@ -436,7 +281,7 @@ HRESULT FDXDShaderManager::AddVertexShaderAndInputLayout(const std::wstring& Fil
 
     VertexShaderCSO->Release();
 
-    std::size_t currentHash = ComputeFileHash(FileName);
+    std::size_t currentHash = ShaderHashUtils::ComputeFileHash(FileName);
 
     FShaderReloadInfo reloadInfo;
     reloadInfo.FileName = FileName;
@@ -487,96 +332,4 @@ ID3D11ComputeShader* FDXDShaderManager::GetComputeShaderByKey(const std::wstring
         return *ComputeShaders.Find(Key);
     }
     return nullptr;
-}
-
-// 파일 내용의 해시값을 사용해 변경 여부를 판단하는 핫리로드 함수
-void FDXDShaderManager::CheckAndReloadShaders()
-{
-    for (auto& pair : ShaderReloadMap)
-    {
-        FShaderReloadInfo& info = pair.Value;
-
-        // 현재 파일 해시값 계산
-        std::size_t currentHash = ComputeFileHash(info.FileName);
-        // 해시값이 변경되었으면 재컴파일 수행
-        if (currentHash != info.FileHash)
-        {
-            UINT shaderFlags = D3DCOMPILE_ENABLE_STRICTNESS;
-#ifdef _DEBUG
-            shaderFlags |= D3DCOMPILE_DEBUG | D3DCOMPILE_SKIP_OPTIMIZATION;
-#endif
-
-            HRESULT hr = S_OK;
-            if (info.ShaderType == EShaderType::Vertex)
-            {
-
-                ID3DBlob* VsBlob = nullptr;
-                const D3D_SHADER_MACRO* Defines = GetShaderMacro(info.ViewMode);
-                hr = CompileShaderWithTempFile(info.FileName, Defines, info.EntryPoint, "vs_5_0", shaderFlags, &VsBlob);
-                if (SUCCEEDED(hr))
-                {
-                    ID3D11VertexShader* NewVertexShader = nullptr;
-                    hr = DXDDevice->CreateVertexShader(VsBlob->GetBufferPointer(), VsBlob->GetBufferSize(), nullptr, &NewVertexShader);
-                    if (SUCCEEDED(hr))
-                    {
-                        // 기존 버텍스 셰이더 해제
-                        if (VertexShaders.Contains(info.ShaderKey))
-                        {
-                            VertexShaders[info.ShaderKey]->Release();
-                            VertexShaders.Remove(info.ShaderKey);
-                        }
-                        VertexShaders[info.ShaderKey] = NewVertexShader;
-                        info.FileHash = currentHash;
-                        GraphicDevice->DeviceContext->VSSetShader(NewVertexShader, nullptr, 0);
-                    }
-                    else
-                    {
-                        MessageBox(nullptr, L"Failed to Create Vertex Shader!", L"Error", MB_ICONERROR | MB_OK);
-                    }
-                    VsBlob->Release();
-                }
-                else
-                {
-                    MessageBox(nullptr, L"Failed to Compile Vertex Shader!", L"Error", MB_ICONERROR | MB_OK);
-                    OutputDebugStringA("Vertex Shader 재컴파일 실패\n");
-                }
-            }
-            else if (info.ShaderType == EShaderType::Pixel)
-            {
-
-                ID3DBlob* PsBlob = nullptr;
-                const D3D_SHADER_MACRO* Defines = GetShaderMacro(info.ViewMode);
-                hr = CompileShaderWithTempFile(info.FileName, Defines, info.EntryPoint, "ps_5_0", shaderFlags, &PsBlob);
-
-                if (SUCCEEDED(hr))
-                {
-                    ID3D11PixelShader* NewPixelShader = nullptr;
-                    hr = DXDDevice->CreatePixelShader(PsBlob->GetBufferPointer(), PsBlob->GetBufferSize(), nullptr, &NewPixelShader);
-                    if (SUCCEEDED(hr) && NewPixelShader)
-                    {
-
-                        // 기존 픽셀 셰이더 해제
-                        if (PixelShaders.Contains(info.ShaderKey))
-                        {
-                            PixelShaders[info.ShaderKey]->Release();
-                            PixelShaders.Remove(info.ShaderKey);
-                        }
-
-                        PixelShaders[info.ShaderKey] = NewPixelShader;
-                        info.FileHash = currentHash;
-                    }
-                    else
-                    {
-                        MessageBox(nullptr, L"Failed to Create Pixel Shader!", L"Error", MB_ICONERROR | MB_OK);
-                    }
-                    PsBlob->Release();
-                }
-                else
-                {
-                    MessageBox(nullptr, L"Failed to Compile Pixel Shader!", L"Error", MB_ICONERROR | MB_OK);
-                    OutputDebugStringA("Pixel Shader 재컴파일 실패\n");
-                }
-            }
-        }
-    }
 }
